03.baselines.demo/3D V-net/requirements.txt
pytorch=1.5
scipy=1.4.1
numpy=1.18
pillow=7.1.2
ipython=7.14
matplotlib=3.2.1
networkx=2.4
pandas=1.0.3